# back/internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ExpenseRequest defines model for ExpenseRequest.
type ExpenseRequest struct {
	Amount    int       `json:"amount"`
	Category  string    `json:"category"`
	Date      time.Time `json:"date"`
	Memo      *string   `json:"memo,omitempty"`
	StoreName string    `json:"store_name"`
	UserId    int       `json:"user_id"`
}

// ExpenseResponse defines model for ExpenseResponse.
type ExpenseResponse struct {
	Amount    int       `json:"amount"`
	Category  string    `json:"category"`
	CreatedAt time.Time `json:"created_at"`
	Date      time.Time `json:"date"`
	Id        int       `json:"id"`
	Memo      *string   `json:"memo,omitempty"`
	PayerName *string   `json:"payer_name,omitempty"`
	StoreName string    `json:"store_name"`
	UserId    int       `json:"user_id"`
}

// LinkAccountRequest defines model for LinkAccountRequest.
type LinkAccountRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// SignUpRequest defines model for SignUpRequest.
type SignUpRequest struct {
	Email    openapi_types.Email `json:"email"`
	Image    *string             `json:"image,omitempty"`
	Name     string              `json:"name"`
	Password string              `json:"password"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	Admin     bool                 `json:"admin"`
	CreatedAt time.Time            `json:"created_at"`
	Email     *openapi_types.Email `json:"email"`
	Id        int                  `json:"id"`
	Image     *string              `json:"image,omitempty"`
	Name      string               `json:"name"`
}

// UserUpdate defines model for UserUpdate.
type UserUpdate struct {
	Image *string `json:"image,omitempty"`
	Name  *string `json:"name,omitempty"`
}

// GetApiV1AuthLineCallbackParams defines parameters for GetApiV1AuthLineCallback.
type GetApiV1AuthLineCallbackParams struct {
	// Code Authorization code from LINE
	Code string `form:"code" json:"code"`

	// State State parameter for CSRF protection
	State string `form:"state" json:"state"`
}

// GetExpensesParams defines parameters for GetExpenses.
type GetExpensesParams struct {
	// Year Year to filter expenses
	Year int `form:"year" json:"year"`

	// Month Month to filter expenses
	Month int `form:"month" json:"month"`

	// Category Category to filter expenses
	Category *string `form:"category,omitempty" json:"category,omitempty"`
}

// PostApiV1AuthLineLinkJSONRequestBody defines body for PostApiV1AuthLineLink for application/json ContentType.
type PostApiV1AuthLineLinkJSONRequestBody = LinkAccountRequest

// PostExpensesJSONRequestBody defines body for PostExpenses for application/json ContentType.
type PostExpensesJSONRequestBody = ExpenseRequest

// PutExpensesIdJSONRequestBody defines body for PutExpensesId for application/json ContentType.
type PutExpensesIdJSONRequestBody = ExpenseRequest

// PostSignupJSONRequestBody defines body for PostSignup for application/json ContentType.
type PostSignupJSONRequestBody = SignUpRequest

// PutUserJSONRequestBody defines body for PutUser for application/json ContentType.
type PutUserJSONRequestBody = UserUpdate