# back/openapi.yaml
openapi: 3.0.3
info:
  title: Example API
  version: 1.0.0
paths:
  /signup:
    post:
      tags:
        - user
      summary: User registration
      description: Register a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /api/v1/auth/line/login:
    get:
      tags:
        - auth
      summary: Initiate LINE login flow
      description: Returns the LINE authorization URL.
      responses:
        '200':
          description: Successfully returned LINE authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string
                    format: uri
                    description: LINE authorization URL
        '500':
          description: Internal server error
  /api/v1/auth/line/callback:
    get:
      tags:
        - auth
      summary: LINE login callback
      description: Handles the callback from LINE. If user exists, logs in. If not, returns unregistered status.
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
          description: Authorization code from LINE
        - in: query
          name: state
          schema:
            type: string
          required: true
          description: State parameter for CSRF protection
      responses:
        '200':
          description: LINE login successful. Sets JWT cookie.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
                    enum: ["logged_in", "unregistered"]
                  line_name:
                    type: string
                    description: Returned only if status is unregistered
                  line_picture:
                    type: string
                    description: Returned only if status is unregistered
        '400':
          description: Invalid code or state
        '401':
          description: Invalid CSRF state
        '500':
          description: Internal server error
  /api/v1/auth/line/link:
    post:
      tags:
        - auth
      summary: Link LINE account to existing user
      description: Links a pending LINE account (from pre-auth cookie) to an existing email/password account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkAccountRequest'
      responses:
        '200':
          description: Account linked and logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input or missing pre-auth cookie
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error
  /api/v1/auth/line/create:
    post:
      tags:
        - auth
      summary: Create new user from LINE account
      description: Creates a new user using the pending LINE account information (from pre-auth cookie).
      responses:
        '201':
          description: User created and logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Missing pre-auth cookie
        '500':
          description: Internal server error
  /user:
    put:
      tags:
        - user
      summary: Update user information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /expenses:
    post:
      tags:
        - expense
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
    get:
      tags:
        - expense
      summary: Get expenses
      parameters:
        - in: query
          name: year
          schema:
            type: integer
          required: true
          description: Year to filter expenses
        - in: query
          name: month
          schema:
            type: integer
          required: true
          description: Month to filter expenses
        - in: query
          name: category
          schema:
            type: string
          required: false
          description: Category to filter expenses
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseResponse'
        '500':
          description: Internal server error
  /expenses/{id}:
    put:
      tags:
        - expense
      summary: Update an expense
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the expense to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '200':
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          description: Invalid input
        '404':
          description: Expense not found
        '500':
          description: Internal server error
    delete:
      tags:
        - expense
      summary: Delete an expense
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the expense to delete
      responses:
        '204':
          description: Expense deleted successfully
        '404':
          description: Expense not found
        '500':
          description: Internal server error
components:
  schemas:
    SignUpRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        image:
          type: string
    LinkAccountRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    UserResponse:
      type: object
      required:
        - id
        - name
        - admin
        - created_at
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
          nullable: true
        name:
          type: string
        image:
          type: string
        admin:
          type: boolean
        created_at:
          type: string
          format: date-time
    ExpenseResponse:
      type: object
      required:
        - id
        - user_id
        - amount
        - store_name
        - date
        - category
        - created_at
      properties:
        id:
          type: integer
        user_id:
          type: integer
        amount:
          type: integer
        store_name:
          type: string
        date:
          type: string
          format: date-time
        category:
          type: string
        memo:
          type: string
        created_at:
          type: string
          format: date-time
        payer_name:
          type: string

    UserUpdate:
      type: object
      properties:
        name:
          type: string
        image:
          type: string
    ExpenseRequest:
      type: object
      required:
        - amount
        - store_name
        - date
        - category
        - user_id
      properties:
        amount:
          type: integer
        store_name:
          type: string
        date:
          type: string
          format: date-time
        category:
          type: string
        memo:
          type: string
        user_id:
          type: integer